pyproject.toml
[project]
name = "ds-rpc-01"
version = "0.1.0"
description = "Starter project for the RPC-01: Internal Chatbot with Role Based Access Control"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn[standard]",
    "langchain-community",
    "chromadb",
    "sentence-transformers",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv",
    "requests",
    "tabulate"
]
